refactor: consolidate 'typings' module into 'code-generator'
There's no point in keeping this module separate, now that we moved the
typings to github-actions-typing-catalog.
