+++ automation/code-generator/build.gradle.kts
@@ -17,7 +17,6 @@ dependencies {
     implementation("io.ktor:ktor-serialization-kotlinx-json")
     implementation("io.arrow-kt:arrow-fx-coroutines:1.2.1")
 
-    implementation(projects.automation.typings)
     implementation(projects.actionBindingGenerator)
 
     testImplementation(projects.githubWorkflowsKt)
