+++ automation/code-generator/src/main/kotlin/io/github/typesafegithub/workflows/codegenerator/ActionBindingsToGenerate.kt
@@ -1,0 +1,3 @@
+package io.github.typesafegithub.workflows.codegenerator
+
+val bindingsToGenerate by lazy { readActionsMetadata() }
