+++ automation/code-generator/src/main/kotlin/io/github/typesafegithub/workflows/codegenerator/model/ActionBindingRequest.kt
@@ -1,4 +1,4 @@
-package io.github.typesafegithub.workflows.actionsmetadata.model
+package io.github.typesafegithub.workflows.codegenerator.model
 
 import io.github.typesafegithub.workflows.actionbindinggenerator.ActionCoords
