--- automation/typings/build.gradle.kts
@@ -1,7 +1,0 @@
-plugins {
-    buildsrc.convention.`kotlin-jvm`
-}
-
-dependencies {
-    implementation(projects.actionBindingGenerator)
-}
