--- automation/typings/src/main/kotlin/io/github/typesafegithub/workflows/actionsmetadata/ActionBindingsToGenerate.kt
@@ -1,3 +1,0 @@
-package io.github.typesafegithub.workflows.actionsmetadata
-
-val bindingsToGenerate by lazy { readActionsMetadata() }
